moved advertisers-page folder out to pages folder
