+++ packages/p2p/src/app-content.jsx
@@ -3,7 +3,7 @@ import { isMobile } from '@deriv/shared';
 import { Loading, Tabs } from '@deriv/components';
 import { observer } from 'mobx-react-lite';
 import { useStores } from 'Stores';
-import AdvertiserPage from 'Components/advertiser-page/advertiser-page.jsx';
+import AdvertiserPage from 'Pages/advertiser-page';
 import BuySell from 'Components/buy-sell/buy-sell.jsx';
 import Dp2pBlocked from 'Components/dp2p-blocked';
 import { localize } from 'Components/i18next';
