+++ packages/p2p/src/components/my-profile/my-profile-stats/my-profile-name/my-profile-name.jsx
@@ -9,7 +9,7 @@ import TradeBadge from '../../../trade-badge';
 import MyProfilePrivacy from '../my-profile-privacy';
 import StarRating from 'Components/star-rating';
 import RecommendedBy from 'Components/recommended-by';
-import BlockUserCount from 'Components/advertiser-page/block-user/block-user-count';
+import BlockUserCount from 'Pages/advertiser-page/block-user/block-user-count';
 
 const MyProfileName = () => {
     const { general_store, my_profile_store } = useStores();
