+++ packages/p2p/src/pages/advertiser-page/advertiser-page-adverts/advertiser-page-row/index.js
@@ -1,0 +1,3 @@
+import AdvertiserPageRow from './advertiser-page-row.jsx';
+
+export default AdvertiserPageRow;
