+++ packages/p2p/src/pages/advertiser-page/advertiser-page-adverts/index.js
@@ -1,0 +1,3 @@
+import AdvertiserPageAdverts from './advertiser-page-adverts.jsx';
+
+export default AdvertiserPageAdverts;
