+++ packages/p2p/src/pages/advertiser-page/advertiser-page-dropdown-menu/index.js
@@ -1,0 +1,3 @@
+import AdvertiserPageDropdownMenu from './advertiser-page-dropdown-menu.jsx';
+
+export default AdvertiserPageDropdownMenu;
