+++ packages/p2p/src/pages/advertiser-page/advertiser-page-stats/index.js
@@ -1,0 +1,3 @@
+import AdvertiserPageStats from './advertiser-page-stats';
+
+export default AdvertiserPageStats;
