+++ packages/p2p/src/pages/advertiser-page/advertiser-page.jsx
@@ -11,11 +11,11 @@ import BuySellModal from 'Components/buy-sell/buy-sell-modal.jsx';
 import PageReturn from 'Components/page-return/page-return.jsx';
 import RecommendedBy from 'Components/recommended-by';
 import UserAvatar from 'Components/user/user-avatar/user-avatar.jsx';
-import AdvertiserPageStats from './advertiser-page-stats.jsx';
-import AdvertiserPageAdverts from './advertiser-page-adverts.jsx';
+import AdvertiserPageStats from './advertiser-page-stats';
+import AdvertiserPageAdverts from './advertiser-page-adverts';
 import StarRating from 'Components/star-rating';
-import AdvertiserPageDropdownMenu from './advertiser-page-dropdown-menu.jsx';
-import TradeBadge from '../trade-badge/trade-badge.jsx';
+import AdvertiserPageDropdownMenu from './advertiser-page-dropdown-menu';
+import TradeBadge from 'Components/trade-badge/trade-badge.jsx';
 import BlockUserOverlay from './block-user/block-user-overlay';
 import BlockUserModal from 'Components/block-user/block-user-modal';
 import ErrorModal from 'Components/error-modal/error-modal';
