+++ packages/p2p/src/pages/advertiser-page/advertiser-page.scss
