+++ packages/p2p/src/pages/advertiser-page/block-user/block-user-count/index.js
@@ -1,0 +1,3 @@
+import BlockUserCount from './block-user-count.jsx';
+
+export default BlockUserCount;
