+++ packages/p2p/src/pages/advertiser-page/block-user/block-user-overlay/index.js
