+++ packages/p2p/src/pages/advertiser-page/index.js
@@ -1,0 +1,3 @@
+import AdvertiserPage from './advertiser-page.jsx';
+
+export default AdvertiserPage;
