--- packages/p2p/src/pages/test.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const test = () => {
-    return <div>test</div>;
-};
-
-export default test;
